requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
coreapi-cli==1.0.9
coreschema==0.0.4
cryptography==2.9.2
Django==3.1.1
django-crispy-forms==1.9.2
django-filter==2.3.0
djangorestframework==3.11.1
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pretty-json==1.2.0
pycparser==2.20
Pygments==2.6.1
pygments-json==0.1
PyJWT==1.7.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.10
